Create README.md
Description
They are colorful, stylish, and versatile characters whose futures depend on a DAO of NFT/ Token holders. Each Optimistic Trash Panda owner will receive an airdrop of Governance Tokens which will enable each owner to have a stake in the future of the collection's collectible characters.

The only details currently set in stone:

-Initial mint will be publicaly available through Optimism.

-All owners on an undisclosed date will receive TxP governance tokens split evenly and allocated based on # of Trash Panda's you own.

-The collection will be interoperable meaning able to travel to other blockchains from the get go.

Upon end of public sale and after the airdrop the artist TrashxPanda will only hold a 10% stake in the tokens and the rest will be distributed. Any tokens that go undistributed and any NFTs that go unsold will be burned.

The artist wants people to make this their own and expand the Universe for the characters to everything from merch, to video games and animated narratives. These characters are a gift to the world to expand upon and make it their own. The DAO can even set up a corporate structure should they be successful in marketing their original ventures. Profits would be distributed based on token share held at the end of each quarter.

-The sky is the limit.